Renamed ProtocolBridge namespace to Factory
