+++ src/CompletionProvider.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 namespace LanguageServer;
 
 use LanguageServer\Index\ReadableIndex;
-use LanguageServer\ProtocolBridge\CompletionItemFactory;
+use LanguageServer\Factory\CompletionItemFactory;
 use LanguageServer\Protocol\{
     TextEdit,
     Range,
