+++ src/Factory/RangeFactory.php
@@ -1,22 +1,21 @@
 <?php
 
-namespace LanguageServer\ProtocolBridge;
+namespace LanguageServer\Factory;
 
-use LanguageServer\Protocol\Location;
 use LanguageServer\Protocol\Position;
 use LanguageServer\Protocol\Range;
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\PositionUtilities;
 
-class LocationFactory
+class RangeFactory
 {
     /**
-     * Returns the location of the node
+     * Returns the range the node spans
      *
      * @param Node $node
      * @return self
      */
-    public static function fromNode(Node $node): Location
+    public static function fromNode(Node $node)
     {
         $range = PositionUtilities::getRangeFromPosition(
             $node->getStart(),
@@ -24,9 +23,9 @@ class LocationFactory
             $node->getFileContents()
         );
 
-        return new Location($node->getUri(), new Range(
+        return new Range(
             new Position($range->start->line, $range->start->character),
             new Position($range->end->line, $range->end->character)
-        ));
+        );
     }
 }
