+++ src/TreeAnalyzer.php
@@ -3,7 +3,7 @@ declare(strict_types = 1);
 
 namespace LanguageServer;
 
-use LanguageServer\ProtocolBridge\RangeFactory;
+use LanguageServer\Factory\RangeFactory;
 use LanguageServer\Protocol\{Diagnostic, DiagnosticSeverity, Range, Position};
 use phpDocumentor\Reflection\DocBlockFactory;
 use Microsoft\PhpParser;
